# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Eco-Chamber",
    "description": "",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
    "/api/categories": {
      "get": {
        "tags": ["categories"],
        "responses": {}
      }
    },
    "/api/events": {
      "get": {
        "tags": ["events"],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/api/events/articles": {
      "get": {
        "tags": ["events"],
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/api/events/search": {
      "get": {
        "tags": ["events"],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/topEvents": {
      "get": {
        "tags": ["topEvents"],
        "responses": {}
      }
    },
    "/api/events/single-sided": {
      "get": {
        "tags": ["events"],
        "parameters": [
          {
            "name": "bias",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
     "/api/gameEvents": {
      "get": {
        "tags": ["gameEvents"],
        "responses": {}
      }
    },
    "/api/eventSentiment": {
      "get": {
        "tags": ["eventSentiment"],
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/api/sources": {
      "get": {
        "tags": ["sources"],
        "parameters": [
          {
            "name": "eventId",
            "in": "query",
            "type": "integer"
          }
        ],
        "responses": {}
      }
    },
    "/api/users": {
      "post": {
        "tags": ["users"],
        "parameters": [
          {
            "name": "UserData",
            "in": "body",
            "required": "true",
            "schema": {
              "$ref": "#/definitions/UserData"
            }
          }
        ],
        "responses": {}
      }
    },
    "/api/auth/login": {
      "get": {
        "tags": ["auth"],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
    "UserData": {
      "properties": {
        "email": {
          "type": "string",
          "required": "true"
        },
        "password": {
          "type": "string",
          "required": "true"
        }
      }
    }
  }
}